requirements.txt
torch
pyyaml
numpy
matplotlib
pandas
gitpython
